updating tsconfig and texttosql
