+++ examples/hacker-news-analyzer/gensx.yaml
@@ -1,0 +1,1 @@
+projectName: "hacker-news-analyzer"
