+++ examples/vercel-ai/package.json
@@ -9,23 +9,20 @@
   "description": "Example of using GenSX with the Vercel AI SDK",
   "scripts": {
     "dev": "pnpm build && node dist/index.js",
-    "start": "npx gensx start ./src/workflows.tsx",
+    "start": "npx gensx start ./src/workflows.ts",
     "build": "tsc -p tsconfig.build.json",
-    "deploy": "npx gensx deploy -e OPENAI_API_KEY ./src/workflows.tsx",
+    "deploy": "npx gensx deploy -e OPENAI_API_KEY ./src/workflows.ts",
     "prepare": "ts-patch install -s"
   },
   "dependencies": {
     "@ai-sdk/openai": "^1.3.22",
-    "@gensx/core": "0.4.0-alpha.3",
-    "@gensx/vercel-ai": "0.2.0-alpha2",
+    "@gensx/core": "workspace:*",
+    "@gensx/vercel-ai": "workspace:*",
     "ai": "^4.2.10",
-    "openai": "^4.87.4",
     "zod": "^3.24.2"
   },
   "devDependencies": {
     "@types/node": "^22",
-    "nodemon": "^3.1.9",
-    "tsx": "^4.19.2",
     "ts-function-decorator": "^0.0.1",
     "ts-function-decorator-ls": "^0.0.1",
     "ts-patch": "^3.3.0",
